+++ apps/docs/theme.config.tsx
@@ -1,0 +1,26 @@
+import React from "react";
+import { DocsThemeConfig } from "nextra-theme-docs";
+
+const config: DocsThemeConfig = {
+  head: (
+    <>
+      <link
+        rel="stylesheet"
+        href="https://fonts.googleapis.com/css?family=Inter:400,500,600,700"
+      />
+    </>
+  ),
+  logo: <span>My Project</span>,
+  project: {
+    link: "https://github.com/shuding/nextra-docs-template",
+  },
+  chat: {
+    link: "https://discord.com",
+  },
+  docsRepositoryBase: "https://github.com/shuding/nextra-docs-template",
+  footer: {
+    text: "Nextra Docs Template",
+  },
+};
+
+export default config;