+++ packages/fpkit-nextjs/src/components/layout/header.tsx
@@ -1,36 +1,76 @@
-import { Header, Title, Button, Text } from "@fpkit/react";
 import React from "react";
-
+import { Title, Text, Tag } from "@fpkit/react";
+import { Header, Section } from "./landmarks";
 /** set types  */
 export interface HeaderProps {
-  title: React.ReactNode;
-  ctaLink?: string;
-  ctaLabel?: React.ReactNode;
+  title?: React.ReactNode;
+  link?: string;
+  linkLabel?: React.ReactNode;
+  description?: React.ReactNode;
   children?: React.ReactNode;
 }
 
 export type BaseProps = { children: React.ReactNode };
-export const Base = ({ children }: BaseProps) => {
+export const HeaderBase = ({ children }: BaseProps) => {
   return <Header>{children}</Header>;
 };
 
-export const ThemeHeader = ({ title, ctaLink, ctaLabel }: HeaderProps) => {
+export function DefaultHeader({
+  title,
+  link,
+  linkLabel,
+  description,
+}: HeaderProps) {
+  return (
+    <Section>
+      <Title
+        elm="h2"
+        styles={{
+          "--fs": "var(--h1)",
+        }}
+      >
+        {title}
+      </Title>
+      {description && <Text elm="p">{description}</Text>}
+      <Text elm="p">
+        <Tag
+          as="a"
+          href={link}
+          data-btn
+          data-variant="primary"
+          data-style="pill"
+        >
+          {linkLabel || "Get Started"}
+        </Tag>
+      </Text>
+    </Section>
+  );
+}
+
+export const ThemeHeader = ({
+  title,
+  link,
+  linkLabel,
+  description,
+  children,
+}: HeaderProps) => {
   if (!title) {
     throw new Error("ThemeHeader requires a title prop");
   }
   return (
-    <Base>
-      <Title elm="h2" styles={{ "--fs": "var(--h1)" }}>
-        {title}
-      </Title>
-      <Text elm="p">
-        <Button type="button" data-variant="primary" data-style="pill">
-          {ctaLabel || "Get Started"}
-        </Button>
-      </Text>
-    </Base>
+    <HeaderBase>
+      {children ? (
+        children
+      ) : (
+        <DefaultHeader
+          title={title}
+          link={link}
+          linkLabel={linkLabel}
+          description={description}
+        />
+      )}
+    </HeaderBase>
   );
 };
-
 ThemeHeader.displayName = "Header";
 export default ThemeHeader;