+++ packages/fpkit-nextjs/src/components/nav/navbar.test.tsx
@@ -10,15 +10,15 @@ describe("Nav renders correctly", () => {
   test("renders without crashing", async () => {
     const navList = [
       {
-        url: "/home",
+        route: "/home",
         name: "Home",
       },
       {
-        url: "/about",
+        route: "/about",
         name: "About",
       },
       {
-        url: "/contact",
+        route: "/contact",
         name: "Contact",
       },
     ];