+++ packages/fpkit-nextjs/src/hooks/useOpts.tsx
@@ -1,26 +1,24 @@
-import { PageOpts, NextraThemeLayoutProps } from "nextra";
+import { NextraThemeLayoutProps, PageMapItem } from "nextra";
 
 export type optsTypes = {
-  options: NextraThemeLayoutProps;
+  options: Pick<NextraThemeLayoutProps, "pageOpts">;
 };
 
 export const usePageOpts = ({ options }: optsTypes) => {
   const { pageOpts } = options;
 
-  const { pageMap, frontMatter } = pageOpts;
+  const { pageMap } = pageOpts;
 
   if (!pageMap) {
     return { contentDirs: [], postList: [], posts: [] };
   }
 
   const postList = pageMap.filter(
-    (item) =>
+    (item: PageMapItem) =>
       item.kind === "MdxPage" &&
       item.route !== "/" &&
       item.frontMatter !== undefined
   );
-
-  // const posts = postList.filter((item) => item?.frontMatter === true);
 
   const dirList = pageMap.filter(
     (item: { kind: string }) => item.kind === "Folder"