+++ packages/fpkit-nextjs/src/index.ts
@@ -1,6 +1,8 @@
-export { Navbar } from "./components/nav/navbar";
+
 export * from './components/nav/nav';
+export { Navigation } from './components/nav/navigation';
 export * from './components/layout/main';
 export * from './components/layout/header';
 export * from './components/layout/footer';
 export * from './components/layout/landmarks';
+export * from './components/posts/post-list'