+++ packages/fpkit-nextjs/tsup.config.cjs
@@ -1,10 +1,14 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: ["src/index.ts", "src/hooks.ts", "src/libs.ts"],
+  outDir: "libs",
   splitting: true,
   sourcemap: true,
+  treeshake: true,
+  external: ["react", "react-dom", "next", "nextra", "@fpkit/react"],
   clean: true,
   format: ["esm", "cjs"],
   minify: true,
+  dts: true,
 });