+++ CHANGELOG.md
@@ -3,6 +3,76 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.13](https://github.com/shawn-sandy/fpkit-nextjs/compare/v0.0.12...v0.0.13) (2023-03-31)
+
+**Note:** Version bump only for package fpkit-vite
+
+## [0.0.12](https://github.com/shawn-sandy/fpkit-nextjs/compare/v0.0.11...v0.0.12) (2023-03-31)
+
+**Note:** Version bump only for package fpkit-vite
+
+## [0.0.11](https://github.com/shawn-sandy/fpkit-nextjs/compare/v0.0.10...v0.0.11) (2023-03-31)
+
+### Features
+
+- 📖 add stories for navigation component ([4e0a2c6](https://github.com/shawn-sandy/fpkit-nextjs/commit/4e0a2c6959c6f56a85e422c89421e46d929be32c))
+- add docs app (nextra-docs-template) ([10e003b](https://github.com/shawn-sandy/fpkit-nextjs/commit/10e003bc531913a631f47a880a17d6153c886f06))
+
+## [0.0.10](https://github.com/shawn-sandy/fpkit-nextjs/compare/v0.0.9...v0.0.10) (2023-03-31)
+
+**Note:** Version bump only for package fpkit-vite
+
+## [0.0.9](https://github.com/shawn-sandy/fpkit-nextjs/compare/v0.0.8...v0.0.9) (2023-03-31)
+
+### Features
+
+- ♻️ change library components to MDXComponents ([1bbb477](https://github.com/shawn-sandy/fpkit-nextjs/commit/1bbb477ea512aca438300b34c5ce15262e46d587))
+- 🌟 add library for render mdx components ([b61280b](https://github.com/shawn-sandy/fpkit-nextjs/commit/b61280b6df3ff0e8f4b95b582528ee06a6ef367a))
+- 🌟 add sort option to filter-mdx-pages ([e0429bb](https://github.com/shawn-sandy/fpkit-nextjs/commit/e0429bb116c0e09fd44e82b394997db9179ced98))
+- 🎊 add react support to astro-demo ([1b21fd3](https://github.com/shawn-sandy/fpkit-nextjs/commit/1b21fd39490df5b42d4efef608246e6c7695b910))
+- 🏗️ add astro demo to /apps ([30b54f3](https://github.com/shawn-sandy/fpkit-nextjs/commit/30b54f3450cea13f136904f51e08fdc8b60a1ec9))
+- 🏗️ add filter-mdx-pages ([6e29118](https://github.com/shawn-sandy/fpkit-nextjs/commit/6e2911861491cca9fbf4f2b20967992dbd238763))
+- 🏗️ add tsconfig.json to nextra-blog ([81663a3](https://github.com/shawn-sandy/fpkit-nextjs/commit/81663a30579b95d18af8f35b257b7e6e8f70876d))
+- 🏗️ export all types/interfaces from fpkit ([733d814](https://github.com/shawn-sandy/fpkit-nextjs/commit/733d81438c0fd2fb05406e467cd80bf7d78a9040))
+- 🔥 remove unused code/variables ([94ed4c8](https://github.com/shawn-sandy/fpkit-nextjs/commit/94ed4c80c982cd6b947c702e9e63751419ac0877))
+- 🚗 add post list page ([2523af9](https://github.com/shawn-sandy/fpkit-nextjs/commit/2523af9fcd0a69e700bd67184ee3aa10a8181500))
+- 🚧 add post list component ([eb88019](https://github.com/shawn-sandy/fpkit-nextjs/commit/eb880196a062afb4c81c5dd5848405eb3f4152ea))
+- 🚧 add theme context provider ([7ddf58a](https://github.com/shawn-sandy/fpkit-nextjs/commit/7ddf58a5feeb2c420ff148b9fb8f654b486a9944))
+- 🚧 modify the header components ([0bf5642](https://github.com/shawn-sandy/fpkit-nextjs/commit/0bf5642016776828f3c2b0f03443b17c08a6bff4))
+- 🚧 rename types headerProps ([e1bdf78](https://github.com/shawn-sandy/fpkit-nextjs/commit/e1bdf7811a9f27a6621c9282817533377fe594fc))
+- export filterMdxPages types ([49dd965](https://github.com/shawn-sandy/fpkit-nextjs/commit/49dd965e6e75896fb0f7452f4e11de380f117347))
+
+### BREAKING CHANGES
+
+- 🧨 filter-mdx-pages
+- 🧨 rename types headerProps
+
+## [0.0.8](https://github.com/shawn-sandy/fpkit-nextjs/compare/v0.0.7...v0.0.8) (2023-03-27)
+
+### Features
+
+- 🌟 update theme config/opts ([f262762](https://github.com/shawn-sandy/fpkit-nextjs/commit/f262762bb0c40998a0081be1eb05126db729c884))
+
+## [0.0.7](https://github.com/shawn-sandy/fpkit-nextjs/compare/v0.0.6...v0.0.7) (2023-03-27)
+
+### Features
+
+- modify useOpts hook ([e318d4e](https://github.com/shawn-sandy/fpkit-nextjs/commit/e318d4e2e4a2667f22c52534879389b0910a004c))
+
+## [0.0.6](https://github.com/shawn-sandy/fpkit-nextjs/compare/v0.0.5...v0.0.6) (2023-03-27)
+
+**Note:** Version bump only for package fpkit-vite
+
+## [0.0.5](https://github.com/shawn-sandy/fpkit-nextjs/compare/v0.0.4...v0.0.5) (2023-03-27)
+
+### Features
+
+- 🧑‍🏭 remove usePageOpts hook from index.js ([e24900e](https://github.com/shawn-sandy/fpkit-nextjs/commit/e24900e82d67304e0b450c70db318bd5f6540c73))
+- 🪶 (fpkit-nextjs) add useFpkit hook ([f619845](https://github.com/shawn-sandy/fpkit-nextjs/commit/f619845db2b46c84ea0a9d2cfb0d91bbca2bd807))
+- 🚧 modify add styles to navbar ([308f1cc](https://github.com/shawn-sandy/fpkit-nextjs/commit/308f1cc22f6817cfc5bc2ef993e017e8d43ba258))
+- 🚧 update nextra-blog theme ([a8911ea](https://github.com/shawn-sandy/fpkit-nextjs/commit/a8911eadb669b8f3ceca1d616a510a0faad8ca5b))
+- modified nav logo ([84dc5c9](https://github.com/shawn-sandy/fpkit-nextjs/commit/84dc5c9120bff30a5918f2546b1f17d8f67e3415))
+
 ## [0.0.4](https://github.com/shawn-sandy/fpkit-nextjs/compare/v0.0.3...v0.0.4) (2023-03-26)
 
 ### Features